+++ perf_tests/ze_cabe/src/level-zero/ze_blackscholes.cpp
@@ -143,7 +143,7 @@ template <class T> void ZeBlackScholes<T>::create_cmdlist() {
 template <class T> void ZeBlackScholes<T>::execute_work() {
   ZE_CHECK_RESULT(zeCommandQueueExecuteCommandLists(command_queue, 1,
                                                     &command_list, nullptr));
-  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT32_MAX));
+  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT64_MAX));
 }
 
 template <class T> bool ZeBlackScholes<T>::verify_results() {
