+++ perf_tests/ze_cabe/src/level-zero/ze_simpleadd.cpp
@@ -100,7 +100,7 @@ void ZeSimpleAdd::create_cmdlist() {
 void ZeSimpleAdd::execute_work() {
   ZE_CHECK_RESULT(zeCommandQueueExecuteCommandLists(command_queue, 1,
                                                     &command_list, nullptr));
-  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT32_MAX));
+  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT64_MAX));
 }
 
 bool ZeSimpleAdd::verify_results() {
