+++ perf_tests/ze_cabe/src/level-zero/ze_sobel.cpp
@@ -115,7 +115,7 @@ void ZeSobel::create_cmdlist() {
 void ZeSobel::execute_work() {
   ZE_CHECK_RESULT(zeCommandQueueExecuteCommandLists(command_queue, 1,
                                                     &command_list, nullptr));
-  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT32_MAX));
+  ZE_CHECK_RESULT(zeCommandQueueSynchronize(command_queue, UINT64_MAX));
 }
 
 bool ZeSobel::verify_results() {
